Added missing LICENSE file (PIZZA-WARE).
This script is simple (almost trivial), so I initially added no LICENSE
at all. However, in strict terms, this means that no one is allowed to
use it -- at least according to german law as far as I know it.

Made it PIZZA-WARE because GPL is overkill for such a simple script.