Search highlighting: Don't use tput(1), don't suggest "less -r".
Using "tput sgr0" in a GNU screen session which runs in an XTerm (or the
Linux VT) results in garbled output (you see a "^O"). This could be
circumvented by using "less -r".

"less -r", however, results in a lot of garbled output (even for "git
diff"). The only reasonable choice is to use "less -R".

But "less -R", again, results in garbled output inside of screen/xterm.

So. Why bother using tput in the first place? We're talking about simple
colors, no fancy stuff. It's safe to assume that every terminal today
supports this. If your terminal does not, I'd be interested to know
which terminal you're using.