Add housekeeping via "clean_empty"
Similar to what gitodo does (see 162b192 in the other repo).

What I *would like to do* is remove all files that contain only
whitespace. This, however, is too expensive in a shell script. Instead,
I remove files with a size of 0, 1 or 2 bytes. This accounts for
completely empty files and files with one empty line (UNIX or DOS line
endings).